MERGE: -c 43140 /branches/lwidentity-6.0 -> /trunk/Platform
MERGE: -c 43139 /branches/lwidentity-5.3 -> /branches/lwidentity-6.0

Fix bug 10331 - [solaris 9] lsassd segfault in auth stress

Dlopen failed for the ntlm plugin. dlerror was called to retreive the error string, but it returned NULL. The NULL was passed to sprintf, which causes a segfault on Solaris.

The fix is to use empty string instead of NULL.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43142 9fc1c11e-11b2-4e14-94d8-6c5e05775969