MERGE: -c 43271 /branches/lwidentity-6.0 -> /trunk/Platform
MERGE: -c 43270 /branches/lwidentity-5.3 -> /branches/lwidentity-6.0

Fix bug 10358 - authenticating with invalid UTF-8 string as password gives KRB5_CALL_FAILED

If a string with an invalid UTF-8 encoding is passed as the password to LsaAuthenticateUser, the function returns LW_KRB5_CALL_FAILED, which gives no indication of the real failure.

Arlene says krb5_get_init_creds password is correct in returning EINVAL in this case. So the fix is to preconvert the string to verify its encoding before passing it into kerberos.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43273 9fc1c11e-11b2-4e14-94d8-6c5e05775969