MERGE: -c 43353 /branches/lwidentity-6.0 -> /trunk/Platform
MERGE: -c 43351 /branches/lwidentity-5.3 -> /branches/lwidentity-6.0

Fix bug 10297 - eventlogd takes all lion's share of the cpu under auth stress

On my first attempt at optimizing this, I optimized the check to see if the eventlog database needed to be trimmed. That succeeded and allowed the stress to run fast enough to fill up the database to the point it needed to be trimmed. Next the trim call was the bottleneck.

This change optimizes the call to delete the oldest events when the database exceeds the event count limit.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43355 9fc1c11e-11b2-4e14-94d8-6c5e05775969