Merge r43375 to Platform.
C/R: KStemen

Fix for crash on Mac systems in the logonwindow process when prompted to logon again after the screen-saver. The problem is that on Mac, the call to pam_get_item() will fail to return a pConv structure even though the function returned success. I have added extra code to test for the valid pointer before we dereference it for the conv function...


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43378 9fc1c11e-11b2-4e14-94d8-6c5e05775969