MERGE: -c 43443 /branches/lwidentity-6.0 -> /trunk/Platform
MERGE: -c 43442 /branches/lwidentity-5.3 -> /branches/lwidentity-6.0

fix bug 10382 - lsassd leaks memory on Solaris

The ps output on Solaris shows that lsassd continually uses more memory during stress. Valgrind is not ported to Solaris, so it cannot be checked. Totalview only shows a couple of minor leaks, one of them is in this area of code.

This change lets Solaris use the reentrant resolver functions. Hopefully this will fix a memory leak. Either way, this is a good change.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43444 9fc1c11e-11b2-4e14-94d8-6c5e05775969