MERGE: -c 43474 /branches/lwidentity-6.0 -> /trunk/Platform
MERGE: -c 43473 /branches/lwidentity-5.3 -> /branches/lwidentity-6.0

Fix bug 10366 - eventfwdd dies

The event forwarder was bailing out with an error. This change handles a disconnect exception and turns it into an error code. A few error codes are better mapped, and all errors are ignored in the main loop for the forwarder.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43477 9fc1c11e-11b2-4e14-94d8-6c5e05775969