Un-revert the pkinit change, since it now builds on both 32- and 64-bit platforms.
Original log:

MERGE: -c 42182,42256,42366 branches/lwidentity-6.0 --> trunk/Platform

rev 42182 -- Build the pkinit plugin.

rev 42256 -- Fix build break on HP/UX: Use the automake-provided
$(SHLIBEXT) variable rather than hardcoding ".so" as the shared
library extension.

rev 42366 -- Fix mac build break: apparently $(DYNOBJEXT) is the
correct variable to use, not $(SHLIBEXT).  Those two are the same
on most platforms (including linux and HP/UX, where the previous
build problem happened), but on OS X "shared objects", which are
intended to be loaded with dlopen(), are different from "dynamic
libraries", which are intended to be linked into programs.  The
former have a .so extension, while the latter have .dylib.  On OS
X we want to build pkinit.so, not pkinit.dylib.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@43794 9fc1c11e-11b2-4e14-94d8-6c5e05775969