MERGE: -c 44371 /branches/lwidentity-5.3 -> /trunk/Platform
Fix bug 10309 - Users with 'Do not require Kerberos preauthentication' set can't join systems to domains

The PAC is passed through preauthentication. If preauthentication is not used, the user will not get a PAC. Without a PAC, Windows will not know the user's group membership, which is necessary for joining a domain.

The fix is to use PA-ENC-TIMESTAMP on the first ticket request.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@44376 9fc1c11e-11b2-4e14-94d8-6c5e05775969