Merge r44451 to Platform. Fix for bug 10544.
this bug is caused by lsassd running with no configured environment variables to specify the language and other locale specific settings. This Linux systems would pass this setting automatically, on the Mac it seems that launchctl strips these settings and we have to apply them specifically. The change here adds the missing env variables to our launchctl daemon configuration plist file for lwsmd. lwsmd running with the locale settings will pass them on to the other Likewise daeomons that manages.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@44481 9fc1c11e-11b2-4e14-94d8-6c5e05775969