MERGE: -c 44805,44806 /branches/lwidentity-5.3 -> /trunk/Platform
===from 44805===

Fix bug 10569 - bitrock install fails on brand new Solaris 10u8 system

Some versions of Solaris do not come with a version if id that supports the -u or -g options. I changed the install scripts to fallback to perl in this case.

===from 44806===

Accidentally checked in a debug statement with revision 44805.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@44807 9fc1c11e-11b2-4e14-94d8-6c5e05775969