MERGE: -c 45843 /trunk/Enterprise -> /trunk/Platform
Fix a regression caused by revision 45699

Revision 45699 resets the list pointers after removing an entry. The in memory cache was reading those pointers after the item was removed from the list to determine how many items were still remaining.

The fix is to check how many items are in the list before removing the item.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@45844 9fc1c11e-11b2-4e14-94d8-6c5e05775969