MERGE: -c 46533 /branches/Platform-6.0 -> /trunk/Platform
When refreshing the machine tgt, atomically replace the credentials cache.

The credentials cache was being cleared, then filled with the new TGT. That was causing a small window during a ticket refresh where the cache was empty.

The fix is to write to a new file and move it into place.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@46534 9fc1c11e-11b2-4e14-94d8-6c5e05775969