MERGE: -c 46695 /branches/Platform-6.0 -> /trunk/Platform
Fix bug 10788 - [Solaris - INTERNAL] password changes fail

On Solaris, when changing the password of a domain user from the root account, during the auth phase, the pam_passwd_auth.so.1 is returning PAM_SUCCESS before pam_lsass is called. Presumably, pam_passwd_auth.so.1 returns success because the calling user is root.

Skipping pam_lsass during the auth phase skips our prompt for the current password for password changes. During the chauthtok phase, we have a backup prompt for the current password, but the results were only being stored in PAM_OLDAUTHTOK. Due to other oddities in PAM on Solaris, we retrieve the current password from PAM_LSASS_OLDAUTHTOK instead of PAM_OLDAUTHTOK. So the backup prompt did not store the password in the same place we tried to retrieve it from on Solaris. The fix is to store it in both places.

libgssapi_ntlm.so did not link to libgssapi.so. Without that library, libgssapi_nltm had undefined references for several OIDs. On Linux, those oids are silently retrieved from the parent binary (which always links to libgssapi.so). Apprently Solaris is more picky, and it refused to dlopen libgssapi_ntlm.so.

Password changes use NTLM. NTLM is little endian over the network. Our NTLM code was not converting between host and network endianness. This change fixes the endianness issues for the client NTLM code, which is what password changes use.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@46697 9fc1c11e-11b2-4e14-94d8-6c5e05775969