MERGE: -c 46696 /branches/Platform-6.0 -> /trunk/Platform
In the case of an uninstall and reinstall, only prompt the user to run domainjoin if the machine is not joined.

After a regular uninstall (not purge) the join state is left behind on the system. Sometimes the next install will restore the system to joined.

This change makes the installer more friendly by stating the domain you are joined after the install (whether it is a "new" install or upgrade). If you are not joined, the previous message is displayed.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@46698 9fc1c11e-11b2-4e14-94d8-6c5e05775969