MERGE: -c 47301 /branches/Platform-6.0 ->  /trunk/Platform
Fix bug 10828 - domain join fails on aix

In the init-lwsm.sh script, which is used by deamons managed by lwsm, seq was being called on AIX systems that use DHCP. The seq command is not available on AIX, so starting services was failing.

The fix is to remove the dhcp seq block since it already gets run when lwsm starts.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@47302 9fc1c11e-11b2-4e14-94d8-6c5e05775969