MERGE: -c 43961,44082,44083,44090,44714,46566 /users/keithr/smartcard/Platform --> /trunk/Platform
Merge smartcard PAM changes.

CR: Kyle

------------------------------------------------------------------------
r43961 | CORP\keithr | 2010-05-07 13:58:50 -0700 (Fri, 07 May 2010) | 7 lines

Add an extra invocation of our PAM module with the smartcard_prompt option
before any other module that prompts.  If a smartcard is found, it will
prompt for the PIN, otherwise it will prompt for the password and
save it for later modules.

CR: Kyle

------------------------------------------------------------------------
r44082 | CORP\keithr | 2010-05-11 10:17:16 -0700 (Tue, 11 May 2010) | 4 lines

Add PAM configuration options for the list of services that should
check for a smartcard, and the services that, when prompting for the
PIN, should include the user's GECOS value in the prompt.

------------------------------------------------------------------------
r44083 | CORP\keithr | 2010-05-11 10:18:15 -0700 (Tue, 11 May 2010) | 7 lines

Add a "smartcard_prompt" option to the PAM module, which tells it
to check for a smartcard and prompt for the user's PIN if found.
If no smartcard is found, the module will prompt for the regular
password, since on some platforms the later modules can't cope
with a module before them that sometimes prompts for the password
and sometimes doesn't.

------------------------------------------------------------------------
r44090 | CORP\keithr | 2010-05-11 11:55:35 -0700 (Tue, 11 May 2010) | 4 lines

Modify LsaProcessConfig to optionally return the size of string
and multistring values, via a possibly-NULL DWORD pointer in the
LSA_CONFIG structure.

------------------------------------------------------------------------
r44714 | CORP\keithr | 2010-05-20 14:57:59 -0700 (Thu, 20 May 2010) | 15 lines

Incorporate review feedback:

-   Make sure the "pam_lsass.so smartcard_prompt" line only bails
    without prompting for catastrophic problems that are likely
    to prevent the authentication sequence from succeeding (e.g.
    memory allocation failures).
-   Mark the smartcard_prompt line requisite, so that if it fails,
    later modules won't be called.  This is needed because on some
    platforms later modules such as pam_unix don't deal well with
    the possibility that the password will sometimes be there already
    and sometimes not (try_first_pass isn't universally supported)
-   Change LsaGetSmartCardUser to LsaGetSmartCardUserObject,
    returning a PLSA_SECURITY_OBJECT instead of a PLSA_USER_INFO_*.
-   Use LwAllocateStringPrintf instead of snprintf.

------------------------------------------------------------------------
r46566 | CORP\keithr | 2010-06-30 14:00:45 -0700 (Wed, 30 Jun 2010) | 2 lines

Fix compiler warnings.

(domainjoin: r47382)