MERGE: -c 48262 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 48261 /trunk/Enterprise -> /branches/Platform-6.0

Fix bug 9771 - "Prepend default domain name for AD users and groups" policy is not working

The AD_OnlineFindObjectByName function was returning success with a null result when it could not find a user or group. That was skipping the retry with prependded domain name logic in AD_OnlineFindObjectsByName.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@48263 9fc1c11e-11b2-4e14-94d8-6c5e05775969