Fix bbug 10889 - domainjoin fails to configure gdm-autologin.dpkg-new module
Dpkg creates .dpkg-new files for new configuration files that the user has chosen not to overwrite their exsiting config files with. This is similar to .rpmnew on Redhat systems.

Some services such as gdm-autologin require special processing from domainjoin. Domainjoin has a list of such services inside it. The .dpkg-new extension has to be stripped before comparing the name.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@48314 9fc1c11e-11b2-4e14-94d8-6c5e05775969