MERGE: -c 48653 /branches/Platform-6.0 -> /trunk/Platform
Fix bug 10787 - deadlock condition in NTLM client when eventlogd and lsass are accessing the log file

Consider the scenario where lw-eventlog-cli uses NTLM authentication.
1. lw-eventlog-cli uses NTLM authentication over RPC to eventlogd
2. eventlogd calls into lsass to verify the creds
3. lsass sends an RPC to the DC to verify the creds
4. the DC returns the results
5. lsass calls into eventlogd to record the successful login
6. eventlogd has a global lock for the ntlm authentication, and cannot answer the request.

The fix is to log events asynchronously in a separate thread in lsassd. This way it can queue up the event after the authentication, and return the result to eventlog immediately. This change will also improve the authentication speed when eventlogging is enabled.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@48660 9fc1c11e-11b2-4e14-94d8-6c5e05775969