MERGE: -c 48686, 48691 /trunk/Platform -> /branches/Platform-6.0
Merge fixes for bug 10787 since they have been through a QA pass on trunk.

===from 48686===

Reapply for fix bug 10787 - deadlock condition in NTLM client when eventlogd and lsass are accessing the log file

Consider the scenario where lw-eventlog-cli uses NTLM authentication.
1. lw-eventlog-cli uses NTLM authentication over RPC to eventlogd
2. eventlogd calls into lsass to verify the creds
3. lsass sends an RPC to the DC to verify the creds
4. the DC returns the results
5. lsass calls into eventlogd to record the successful login
6. eventlogd has a global lock for the ntlm authentication, and cannot answer the request.

The fix is to log events asynchronously in a separate thread in lsassd. This way it can queue up the event after the authentication, and return the result to eventlog immediately. This change will also improve the authentication speed when eventlogging is enabled.

This time, I tested that lsass will continue to run even if eventlogd is shutdown and lsass is configured to log events. In that case, lsass will buffer up to about 100 events and try to send them at least 3 times before giving up and dropping them. If eventlogd is started at a later point, any undropped events will be sent.

CR Brian

===from 48691===

Fix a hang on shutdown caused by revision 48686

If eventlogging was enabled in lsassd, and it was shutdown while eventlogd was not running, it would hang. This change moves the check of the conditional variables directly before waiting on them.

Also the variables that can be modified by different threads are marked volatile.

CR Danilo

(eventlog: r48821)