MERGE: -c 48887 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 48886 /branches/lwidentity-5.3 -> /branches/Platform-6.0

Partial fix for bug 10973 - memory leak in lsassd

When the DC times out a connection, it often (but not always) sends a notice of disconnect to the client with an error message saying the connection is timed out. When openldap parsed the NOD, it assigns lr to &dummy_lr.

Later inside of try_read1msg, the lr is freed if it is not &dummy_lr. That causes the error message to be leaked if lr is &dummy_lr. The standard lr free function cannot be called on lr in this case because &dummy_lr is on the stack. So the error message has to be freed directly.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@48889 9fc1c11e-11b2-4e14-94d8-6c5e05775969