Fix bug 11081 - Access denied for lw-eventlog-cli as root to 127.0.0.1
lw-eventlog-cli has special logic where if the destination machine is 127.0.0.1, a local unix domain socket is used instead of the normal tcp connection. The eventlogd gives root access to the unix domain socket even if root is not listed in the allowed users list. Revision 46659 broke this by converting the ip address into a fqdn first.

The fix is to revert 46659. That revision has been superceded by a more proper fix in revision 48821. The original change was only made to Platform-6.0, so the revert only needs to be done in that branch.

Log from revision 46659:

FIXED:  Bug 10710 -  *[Ubuntu10.0464bit desktop-INTERNAL]LWISE : When specifying 127.0.0.1 in lw-eventlog-cli tool,eventlogd daemon is getting aborted

REVIEWED BY: Arlene

See also bug #10787 for additional information

(eventlog: r48919)