Fix bug 10786 - users across a 1-way trust cannot login while offline Fix bug 11060 - one-way trusted domain user authentication fails for .local domains
This change adds a kerberos service locator. With it, the likewise copy of krb5 will first directly talk to netlogon before using the affinity file or DNS to look up a KDC.

For one way trusts to .local domains, netlogon directly uses the resolver. Otherwise, krb5 will fail when it tries to go through MDNS configured through nsswitch.

For the offline scenario, the DC for the domain will be affinitized during the online login. During the offline login, the affinitized domain will be returned instead of generating a realm unknown error because the DNS server can't be reached.

CR Jerry


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@49504 9fc1c11e-11b2-4e14-94d8-6c5e05775969