MERGE: -c 50018 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 50017 /branches/lwidentity-5.3 -> /branches/Platform-6.0

Fix bug 11264 - lw-find-user-by-name --level 2 for restricted user generates logged errors

The lw-find-user-by-name tool is calling the check membership function in lsass in order to print the result. Lsass was logging an error whenever the user was not in a group.

The fix is to reduce the log message level to info.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50021 9fc1c11e-11b2-4e14-94d8-6c5e05775969