MERGE: -c 50208 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 50204 /trunk/Enterprise -> /branches/Platform-6.0

Fix bug 11170 - "./lw-find-objects --by-upn" option is not working

The lw-find-objects command does not specify whether a user or group is being looked up. In the by-upn case, lsass tried to look it up as a user, then fell back to looking it up as a group. Since groups cannot have UPNs, an internal error code bubbled out of lsass.

The fix is to not fallback to groups when looking up upns.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50211 9fc1c11e-11b2-4e14-94d8-6c5e05775969