Fix bug 11344 - lsassd - ad-open-provider/batch.c:3961 assertion failed: '0'
An assertion was failing in lsass when a domain sid was looked up over a one way trust. The assert was for the object type being a user or a group.

The fix is to return an error if the type is a domain instead of failing an assert.

Fix written by Danilo
CR by Kyle


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50436 9fc1c11e-11b2-4e14-94d8-6c5e05775969