MERGE: -c 48617 /branches/Platform-6.0 -> /trunk/Platform
Set try_first_pass on the smartcard_prompt module, except on RHEL<=4 which
has a known problem with not clearing a failed password before calling
into pam_authenticate to retry.  This fixes pam applications that pre-set
the username and password data and don't provide a conversation function,
such as the OS X screensaver.

CR: Kyle


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50462 9fc1c11e-11b2-4e14-94d8-6c5e05775969