MERGE: -c 50490 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 50489 /branches/lwidentity-5.3 -> /branches/Platform-6.0

Fix bug 10973 - memory leak in lsassd

Spnego frees the inner context on error duing gss_init_sec_context. This causes gss_delete_sec_context to later fail when it is called on the union_ctx and not free all the memory.

The fix is to not return an error from spnego's delete function if the context is null.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50491 9fc1c11e-11b2-4e14-94d8-6c5e05775969