Fix bug 11375 - Mac domainjoin fails configuring dsplugin
Rob's system had some junk in its directory service search path, and the name of the entry started with 'Likewise - Active Directory'. Domainjoin treated that as the likewise directory service plugin, and did not know how to remove it.

The fix is to match the dsplugin name exactly, not just the first portion.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50632 9fc1c11e-11b2-4e14-94d8-6c5e05775969