MERGE: -c 51205 /branches/Platform-6.0 -> /trunk/Platform
Fix a SIGSEGV when lsassd is started without lwiod

Without lwio, the dependency check in lsass fails and it immediately exits. In the exit path, lsass would attempt to stop the event log thread, but it was never started.

This fix checks whether the writerThread is set before trying to join it.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51206 9fc1c11e-11b2-4e14-94d8-6c5e05775969