MERGE: -c 51438 /branches/Platform-6.0 -> /trunk/Platform
MERGE: -c 51436 /branches/lwidentity-5.3 -> /branches/Platform-6.0

Fix bug 9704 - DC locator picking child domain to service root domain on 389

Netlogon will sometimes affinitize to a GC in a child domain. Netlogon was mangling the data in the process of saving to the cache and reloading it. That caused the GC for the forest to show up as a DC for the root domain.

CR Danilo


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51439 9fc1c11e-11b2-4e14-94d8-6c5e05775969