Fix bug 11551 - resumable error in pam configuration on CentOS 5.5
Previously, domainjoin would modify /etc/pam.d/runuser from:
auth            sufficient      pam_rootok.so

to:
auth            sufficient      pam_rootok.so
auth            sufficient      pam_lsass.so
auth            required        pam_deny.so

That introduced a prompting pam module to the configuration. On the next pass, domainjoin would try to add a smart card prompt at the beginning of the file.

The runuser program explicitly should not prompt for a password (and only accept root). The fix is to add a special exception for runuser. Now the file is changed to:
auth            sufficient      pam_rootok.so
auth            required        pam_deny.so

CR Keith


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51472 9fc1c11e-11b2-4e14-94d8-6c5e05775969