Fix bug 11569 - we are configuring netsvc.conf hosts line when it's already configured
When domainjoin changes the hostname, it ensures the module that reads /etc/hosts is first in the in nsswitch/netsvc.conf.

Domainjoin saw tab, space, colon, and equals as the only legal separators between modules in netsvc.conf/nsswitch.conf. This change removes equals and colon, but adds comma.

So before the change, 'local,bind4' looked like one module instead of two.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51521 9fc1c11e-11b2-4e14-94d8-6c5e05775969