lwio:srv: lwnet session not enumerating all sessions
lwnet session currently doesn't enumerate more than 40 sessions
(40 sessions is the max number of sessions that can be enumerated in
a single call to srv because of our buffer size).
This is because we never return STATUS_MORE_ENTRIES, when we have a
partially filled buffer and have more data to send.

This patch returns STATUS_MORE_ENTRIES when we have a partially filled
buffer, so that we can get the remaining sessions in the next call
using the resume handle.

Note: lwmsg_data_marshal returns STATUS_END_OF_FILE when it cannot
fit the data in the buffer.

This patch also allows STATUS_MORE_ENTRIES for DEV_CTL IRP types
to be returned by the iomgr so that we don't assert in the iomgr.

Submitted by: Aravind Srinivasan <aravind.srinivasan@isilon.com>
Reviewed by: Sriram Nambakam

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51542 9fc1c11e-11b2-4e14-94d8-6c5e05775969