Fix bug 7126 - lsass won't leave an offline domain
This patch fixes the issue where domainjoin will not leave a domain if it cannot talk to the DNS server.

Changes:
domainjoin
the method to determine the joined short domain name is changed.

lsa rpc server
tolerant to not finding a DC for the domain, and removed an unused variable.

ad provider
allow a leave to take place even if the joined state is unknown.

CR Rafal


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51615 9fc1c11e-11b2-4e14-94d8-6c5e05775969