lwio:rdr: Simplify response tracking
Rather than using a hash table of a secondary structure
to track responses, rdr now just tranfers the op context
from the 'pending send' queue to a 'pending response' queue.
Since responses come back in roughly the same order as requests
are sent, and huge numbers of concurrent requests are uncommon,
the linear traversal of the pending response queue to look up
a mid is probably outweighed by the decreased memory usage
and cache pressure.

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51734 9fc1c11e-11b2-4e14-94d8-6c5e05775969