Fix bug 11387 - while running stress lsass daemon is crashing
The registry pstore backend was using a global variable. That global variable was getting overwritten when multiple handles were outstanding. Most often this causes lsass to crash when a machine password change and a login occur at the same time.

The fix is to move the global variable into the pstore handle.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@51987 9fc1c11e-11b2-4e14-94d8-6c5e05775969