Work around for bug 11679 - domain join fails with error unable to start daemon
Bug 11679 happened because lwsmd was still running when the previous build was uninstalled. The HP-UX package manager renamed lwsm's plugins with a # in front instead of deleting them (because an in-use file cannot be deleted). The new build tried to open the old plugin.

The work around is to have lwsmd ignore plugins that start with #.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@52318 9fc1c11e-11b2-4e14-94d8-6c5e05775969