MERGE: -c 52588 /trunk/Platform -> /branches/Platform-6.0
Fix bug 11715 - gdm-fingerprint service not known by domainjoin

On RHEL 6, the gdm-fingerprint pam service authenticates users by fingerprint, not by password. So pam_lsass should not be included in the auth phase.

Domainjoin was confused because the 'required pam_deny.so' was added through an include and not directly in the service.

(domainjoin: r52589)