MERGE: -c 52633 /trunk/Platform -> /branches/Platform-6.0
Fix bug 11729 - domainjoin is not successful and its crashing

There was a regression in domainjoin that caused it to more often then necessary add a 'required pam_deny.so' to the end of some services. It then crashed when pam_deny.so was not found.

This change fixes the regression in the logic, includes a new path to find pam_deny.so, and makes the code more robust against crashing.

(domainjoin: r52634)