Fix bug 11207 - pam configuration does not fully complete during joins
On the first pass, domainjoin adds pam_lsass to a service that previously did not prompt for a password (and used something like pam_console before). On the second pass, domainjoin sees pam_lsass which prompts, and then complains because the smart card module is missing.

The patch changes the logic to not enable services in the auth phase that did not prompt for a password before.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@52822 9fc1c11e-11b2-4e14-94d8-6c5e05775969