lwmsg: fix regression where idle clients were not timed out
lwmsg uses the 'message' timeout value when a send or recv
is in progress with the client, otherwise it uses the 'idle'
timeout.  However, since the introduction of message multiplexing,
the server is constantly trying to receive the next message
from the client, so the idle timeout was never used.

This change makes the server distinguish between the first
attempt to receive a message (use idle timeout) and subsequent
attemps to receive the rest of it (use message timeout).  It
also adds a regression test.

CR: Kyle

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@53111 9fc1c11e-11b2-4e14-94d8-6c5e05775969