lwio:smbcommon: fix deadlock with lsassd
When establishing an SMB session through rdr using a krb5 TGT,
gss-api still attempts to acquire credentials through the gss_ntlm
provider, which makes an IPC call to lsassd.  If the original
caller was lsassd (in this case, the AuthUserEx path), this
can exceed the worker thread limit and cause a deadlock.

This change explicitly acquires krb5 credentials before calling
gss-api, which avoids the call into the ntlm provider and avoids
the deadlock.

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@53293 9fc1c11e-11b2-4e14-94d8-6c5e05775969