MERGE: -c 53564 /branches/likewise-oem-lexmark -> /trunk/Platform
Add an optional parameter to specify the kerberos credentials cache
to use.  In addition to using the cache for all GSS calls, get
the UPN from the cache, and use it to specify the user that the
certificate is for, and to find the appropriate domain controllers
to talk to.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@54115 9fc1c11e-11b2-4e14-94d8-6c5e05775969