CID: 16670: SECURE_CODING
Coverity reported use of sprintf() as not secure. However, 
the code using sprintf() was not correct. This had to do with
parsing a NetBIOS "Level 2" name back to a string, when the
name included the "scope ID", which is similar to a subdomain
name in DNS. This change fixes the sprintf() usage, and also
fixes the NB name decoding. There is a unit test added for this 
to confirm this code works correctly.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@54868 9fc1c11e-11b2-4e14-94d8-6c5e05775969