lwsm: try to restart services a few times if they die
For now, it is hard-coded to try up to 2 times in a 30
second period beginning with the first restart.  It will
also deal with services that don't announce their death
(like lwio drivers) by polling the status of the reverse
dependency set.  This ensures drivers get restarted as well
if lwiod dies.

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@54872 9fc1c11e-11b2-4e14-94d8-6c5e05775969