CID: 13914: TAINTED_STRING
Parameter passed into LWNetKrb5PrintfFile() is actually the format string
for vfprintf() and that data was obtained from fgets(). So there is a danger
this string could cause a crash, if it contains %format specifiers within
the string.

I believe the intent is to write this data as a string into the specified,
file, thus the addition of the "%s" format specifier to this call.

CR: dalmeida


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@54900 9fc1c11e-11b2-4e14-94d8-6c5e05775969