CID: 10900: DEADCODE
The if (!pDcInfo) [line 693] cannot be ever be true, so this is dead code. 
There are other places in the while() loop that can return NERR_DCNotFound, 
so removal of this code is less worrisome. Also changed while (!pDcInfo) 
to for (;;) as it makes the use of pDcInfo and how the loop 
terminates clearer.



git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@54910 9fc1c11e-11b2-4e14-94d8-6c5e05775969