Fix bug 12097 - lsass daemon crashed on changing password with passwd command
The passwd command tends to truncate the default username. On a QA system, the default username was truncated so that the domain name part and backslash were used, but with a blank sam account name. That null user name was passed to printf, which crashed

The fix is to check for null before calling printf.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@55080 9fc1c11e-11b2-4e14-94d8-6c5e05775969