Fix bug 12202 - lsassd never goes online after reboot
If the domain manager thread received a spurious wakeup, it would always push the next wakeup time to at least a minute from the current time. If it kept getting spurious wakeups in less than a minute,it would never transition the domains online.

The fix is to not recompute the wakeup time on spurious wakeups.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@55680 9fc1c11e-11b2-4e14-94d8-6c5e05775969