Fix bugs 12186 - --assumeDefaultDomain option not working 100% during join and 12236 domainjoin resets value of AssumeDefaultDomain to false from true
For bug 12186, after setting the assumeDefaultDomain registry setting, lsass's configuration needs to be refreshed. In 5.3, this was not a problem because lsass was restarted on every join.

For bug 12236, the --assumeDefaultDomain option was present or not-present. When the option is not present, there was confusion about whether the setting should be disabled in the registry, or left at its current state. To remove any ambiguities, I changed it to a yes/no option.

Finally, in testing I noticed the --userDomainPrefix option was only working if an uppercase domain name was used. To improve usability, domainjoin now always converts the parameter to upper case.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@56221 9fc1c11e-11b2-4e14-94d8-6c5e05775969