MERGE: -c 56366 /branches/Platform-6.1 -> /branches/Platform-6.0
MERGE: -c 56365 /trunk/Platform -> /branches/Platform-6.1

Fix bug 12267 - while doing domain leave, getting first time join warning

The nsswitch module in domainjoin was making a change on AIX on a domain leave, and telling the user they should reboot because it was the first join.

The user does not need to reboot after a leave, so the message should not be displayed in that case.

(domainjoin: r56367)