Fix bug 12268 - use unprefixed usernames for SSO with assumedefaultdomain
The auth_to_local rule transforms a UPN into a username. Before the .k5login file is created, this rule allows one username per UPN to login with SSO (with the .k5login, we allow up to 4 different UPN formats per user).

This change strips the DOMAIN\\ prefix off of users in the default domain when assumedefault domain is set. This makes it more likely that a canonically formatted username will work with SSO.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@56374 9fc1c11e-11b2-4e14-94d8-6c5e05775969