Fix bug 12052 - lsass runs out of FDs on Solaris with large processors
Lwmsg uses 2 fds per thread it creates, and by default it creates a thread for each processor. Solaris has an OS limitation where fds larger than 255 cannot be used inside a FILE*. Lwmsg has an exception to create fewer threads to stay below the fd limit, but that exception was being subverted by lsass raising the fd limit to 256.

Now on Solaris the fd limit is set to 256 for lsass through the registry.

CR: Brian


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@56695 9fc1c11e-11b2-4e14-94d8-6c5e05775969