MERGE: -c 56762 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12298 - resumable error processing pam stack in domainjoin

There was a typo in splitting the basename and dirname components of the symlink target. This ended up causing a failure for symlinks in /etc/pam.d that point outside /etc/pam.d. One example of such a file is the vmware tools service.

(domainjoin: r56763)