MERGE: -c 56989 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12326 - domainjoin --nohosts option needs to allow customer to join with no fqdn configured

When a computer's kernel hostname was resolvable, but resolved to a hostname without a dot in the name, that used to be an error that prevented domainjoin. Now it is downgraded to a recommendation to change.

CR Glenn

(domainjoin: r56990)