Fix bug 11819 - uninstall hangs after reboot in Solaris 10 u9 machine
Bug 11819 is about the uninstall hanging. It turns out this is due to mismatched unlocking/locking for enumerations. The pthread_rwlock_unlock function was failing, but the error was ignored. When lsass was requested to shutdown, it could not obtain a write lock.

This happens because the user and group enumeration handles hold a read lock for the duration of their existence. So the lock is acquired when the enumeration starts, and the lock is released when the enumeration ends. Those are triggered by different IPC messages, so often thread 1 acquired the lock and thread 2 attempted to release the lock.

This revision implements a read/write lock system based off a condition variable and mutex.

CR: Scott


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@57246 9fc1c11e-11b2-4e14-94d8-6c5e05775969