Bug 12309 - Lsassd crash
Initialize the OpenSSL thread handling subsystem, so that OpenSSL calls
can be safely called from multiple threads.  Also initialize the parts
of OpenSSL that are needed by the pkinit plugin, to avoid a race condition
between multiple instances of the plugin that attempt to initialize it
if it's not already done.

Manually merged from /branches/Platform-6.0 r57125




git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@57294 9fc1c11e-11b2-4e14-94d8-6c5e05775969