lwmsg: add LWMSG_ENDPOINT_PAIR
It is sometimes necessary to indicate that an endpoint connected
via an existing fd is a socket pair.  This is because getpeereid()
on FreeBSD and other platforms does not work on fds from socketpair().
Using LWMSG_ENDPOINT_PAIR avoids this call and assumes that the peer
has the same uid and gid as the current process.

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@57616 9fc1c11e-11b2-4e14-94d8-6c5e05775969