eventlog: fix logic bug in name resolution logic
Fix a logic bug that would cause eventlog to believe that
it had failed to resolve all the names in its access control
settings when it had actually succeeded.  This would cause
it to make expensive queries to lsass on every operation
because it would not cache the generated security descriptor.

CR: Kyle

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@57979 9fc1c11e-11b2-4e14-94d8-6c5e05775969