Fix bug 12551 - domainjoin fails with duplicate ssh error
Bug 11923 required detecting if multiple versions of ssh were installed on a system, and displaying an error in that case.

Bug 12551 is a regression due to the fix for that bug. A standard HP-UX ssh install has sshd in multiple locations, although they are symlinked.

The patch removes any duplicate paths from the list by checking inodes.

CR Scott


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@58425 9fc1c11e-11b2-4e14-94d8-6c5e05775969