Fix handling of unenabled users in the memory cache
Unenabled users were being added to the UPN and UID hash tables, but not being removed. Also users with an empty UPN were being added to the UPN table.

Adding users with an empty UPN to the UPN table and unenabled users to the UID table causes lots of hash collisions.

CR Arlene


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@58551 9fc1c11e-11b2-4e14-94d8-6c5e05775969