Fix bug 12661 - domainjoin fails if no option for domainjoin-cli command is specified
The new default logging location is /var/log/likewise-join.log. The /var/log directory does not exist on HP-UX, and domain join failed with an ERROR_FILE_NOT_FOUND.

The fix is to try /var/adm/likewise-join.log if /var/log does not exist, and only display a warning if /var/adm does not exist too.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59041 9fc1c11e-11b2-4e14-94d8-6c5e05775969