srv/smb{1,2}: UPSTREAM: Add support for client side caching
This patch adds the Client side caching support for lwio.

This patch also makes sure that the treeconnect phases of both
SMBv1 and SMB2 has the appropriate CSC flag turned on.

In case of SMBv1 it is the optionalSupport parameter in the extended
tree connect response and in SMB2 it is part of the shareFlags.

CR: snambakam@likewise.com
Committer: tim.prout@isilon.com
Patch by Aravind



git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59279 9fc1c11e-11b2-4e14-94d8-6c5e05775969