srv: UPSTREAM: libshareinfo: Bug 72515 - LWIOD: Command "lwnet session" produces error [ERROR_ARITHMET IC_OVERFLOW] when more than 256 sessions are active.
We kept getting ARITHMETIC_OVERFLOW when we try to enumerate more than
256 sessions.

We were trying to marshal the ResumeHandle (UINT32) as UINT8 in the
SessionEnumSpec. This results in ARITHMETIC_OVERFLOW when trying to enumerate
more than 256 sessions. The same is true with file enum as well.

Committer: tim.prouty@isilon.com
Patch by Aravind



git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59282 9fc1c11e-11b2-4e14-94d8-6c5e05775969