smb/smb1, rdr: UPSTREAM: Only return an extened TCon response if one was requested
The Windows clients don't seem to care, but other do (specifically solaris).
Furthermore, this matches what a windows server returns when the extended
bit isn't set.  smbtorture tends not to set the extended bit, so it's a
good client to exercise this path.

Adapted to work with new version of RDR (gcarter@likewise.com)

Committer: tim.prouty@isilon.com



git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59317 9fc1c11e-11b2-4e14-94d8-6c5e05775969