Fix bug 12709 - PASSWORD_MISMATCH under stress
Bug 12709 was caused by the machine password changing while a user was logging in. The DC sent service tickets encrypted with the new machine password, before the machine password version was looked up through ldap and written in the local database. That caused an integrity check failure when decrypting the TGS.

This patch adds an rw lock around changing the machine password. Authenticating users and initializing schannel use the read lock.

CR Danilo


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59441 9fc1c11e-11b2-4e14-94d8-6c5e05775969