MERGE: -c 59591 /branches/Platform-6.0 -> /trunk/Platform
Tell the krb5 library to keep the cache open while we're reading
it, but then close it before we sleep waiting for it to change.
Apparently krb5 keeps a read lock on the file while it has it open,
which gets in the way of the DS plugin attempting to write to it.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@59768 9fc1c11e-11b2-4e14-94d8-6c5e05775969