MERGE: -c 59591 /branches/Platform-6.0 -> /branches/Platform-6.1
Tell the krb5 library to keep the cache open while we're reading
it, but then close it before we sleep waiting for it to change.
Apparently krb5 keeps a read lock on the file while it has it open,
which gets in the way of the DS plugin attempting to write to it.

(lsass: r59769)