MERGE: -c 59940 /trunk/Platform -> /branches/Platform-6.1
Fix bug 11857 - requiremembershipof no longer reports clear log messages

If the user is blocked because they are not in the requiremembershipof list, log a message from syslog.

This logging cannot be put into lsassd because the access list check is called by the lw-find-user-by-name program, so the message would occur too frequently.

CR Danilo

(lsass: r59941)