Fix bug 12856 - cannot change password for local user
The user_unknown flag properly caused pam_lsass to return PAM_IGNORE from the pam_chauthtok function, but only when the PAM_PRELIM_CHECK flag was set. When the function was recalled with PAM_UPDATE_AUTHTOK set, the module failed to find the previous password prompt, and failed.

This change checks whether the user belongs to lsass before looking for the old password.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@60136 9fc1c11e-11b2-4e14-94d8-6c5e05775969