Fix bug 12892 - pam_lsass set_default_repository excludes pam_ldap account auth on Solaris
Solaris has a repository setting to describe where the user comes from (I don't know why it doesn't use nsswitch all the time). We set the repository to "files" so that it looks at nsswitch for password changes on login. Aparently that messes up logins that are coming from ldap, another repository option.

This change only sets the repository in pam when the user is known to lsass.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@60308 9fc1c11e-11b2-4e14-94d8-6c5e05775969