MERGE: -c 60315 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12892 - pam_lsass set_default_repository excludes pam_ldap account auth on Solaris

Solaris has a repository setting to describe where the user comes from (I don't know why it doesn't use nsswitch all the time). The repository is necessary for password changes. With the default nsswitch modules, Solaris can automatically determine the repository.

Solaris does not recognize "lsass", so we set the repository to "files" early in the pam stack, because that happens to work for users coming from /etc/passwd and lsass. Apparently setting the repository to files prevents ldap users from logging in.

This change introduces an "lsass_users_only" option that can be combined with the "set_default_repository" option to only set the repository if the user is coming from lsass. The downside is the repository will still have to be manually set for non-Likewise users.

(lsass: r60317)