MERGE: -c 60717 /trunk/Platform -> /branches/Platform-6.1
This change adds a no_require_membership option for pam_lsass so that the 'require membership of' setting is not enforced during the auth phase of pam. This setting is not on by default.

If you choose to use it, change this in the pam file:
auth        sufficient      /lib/security/pam_lsass.so      try_first_pass
to:
auth        sufficient      /lib/security/pam_lsass.so      try_first_pass no_require_membership

Most of the changes in the patch are to ensure the require-membership-of setting is enforced before checking if the password is expired (which causes the function to bail out with a success code).

CR: Arlene

(lsass: r60718)