MERGE: -c 58904 /trunk/Platform -> /branches/Platform-6.1
Fix for bug #12454.  When computing an NTLMv2 hash use the target name we put in the response message and not the one we got from the challenge message.  Also, allow a NULL domain name in AD_NetlogonAuthenticationUserEx which is required when using UPN style user names.

CR:  kstemen

(lsass: r60787)