MERGE: -c 52 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12719 - Password change fails when anonymous access is blocked to samr

Lsass was attempting to connect to samr to perform a password change first with the user's credentials and then anonymously.

Windows machines also try connecting with the machine credentials. This is important when the user's password is expired (user creds won't work), and anonymous access to samr is blocked.

(lsass: r60912)