Fix bug 12968 - cache lookup fails when 'AssumeDefaultDomain' is turned on and a lookup is done on a user without the domain prepended on their name
If assume default domain is enabled, lsass currently does an LDAP look up against the DC every time. The cache is not working.

More specifically, this happens:
1. Lookup up the user by alias in the cache
2. Lookup up the user by alias through LDAP
3. Lookup up the user with default domain name prepended in cache (usually works)
4. Lookup up the user with default domain name prepended through LDAP

This patch changes the operations to be:
1. Lookup up the user by alias in the cache
2. If found but expired, skip to step 3. If not found, lookup up the user with default domain name prepended in cache (usually works).
3. Lookup up the user by alias through LDAP
4. Lookup up the user with default domain name prepended through LDAP.
If this works, store the prepended result in the cache and remove any expired alias results in the cache.

The patch is against the Enterprise-6.0 branch (that's what GoDaddy is using), but I will check it into Platform first.

CR: Glenn


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@60916 9fc1c11e-11b2-4e14-94d8-6c5e05775969