MERGE: -c 66 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12976 - When a local user logs in, we still hit up AD to get group info for the user

Extend the syntax of /etc/likewise/user-ignore and /etc/likewise/group-ignore to allow a + on a line by itself to mean also include the objects listed in /etc/passwd or /etc/group, respectively.

This way calling initgroups for local users will not contact lsassd.

(lwadvapi: r60925)