MERGE: -c 133 /trunk/Platform -> /branches/Platform-6.1
Additional fix for bug 12976 - When a local user logs in, we still hit up AD to get group info for the user

In addition to allowing '+' on a line by itself in the user ignore file to denote the users in /etc/passwd should be ignored, allow '+<filename>' too. The file must be in the system passwd format.

(lsass: r60992)