MERGE: -c 215 /trunk/Platform -> /branches/Platform-6.1
Fix bug 13018 - password change fails when user has alternate UPN suffix defined in AD

When the user belongs to DOMAIN, but has an alternate UPN suffix of REALM, lsass was attempting to contact REALM. When REALM does not exist in DNS, the password change failed.

The fix is to use samaccount@DOMAIN, even if the user is assigned a different UPN in AD. That matches the behavior used during password checks.

(lsass: r61074)