MERGE: -c 263 /trunk/Platform -> /branches/Platform-6.1
Allow gss ntlm authentication to work when the domain name is not passed in

If the client does not pass in a domain, the DC will search all the domains for the username inside the NetrLogonSamLogonEx call. The SSPI layer needs to convert "" to NULL to allow lsassd to run through that logic.

CR Adam

(lsass: r61122)