MERGE: -c 274 /trunk/Platform -> /branches/Platform-6.1
Fix bug 13037 - password expiry message is not coming for RHEL distros

The LsaValidateUser function was skipping the normal password expiration check and message.

The account phase is run twice because pam_lsass is listed twice in pam.conf. An additional boolean is necessary to track whether the password message was already shown.

(lsass: r61133)