Fix bug 13110 - crash in lsassd/join/join.c
Not all error conditions were being handled in a call to SamrLookupNames. If the named pipe connection to the DC was reset in the middle of the SamrLookupNames call, lsass would crash.

The error mapping for STATUS_CONNECTION_RESET was not helpful, so it is switched so a better error message is returned from domainjoin.

CR Glenn


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61515 9fc1c11e-11b2-4e14-94d8-6c5e05775969