MERGE: -c 61515 /trunk/Platform -> /branches/Platform-6.1
Fix bug 13110 - crash in lsassd/join/join.c

Not all error conditions were being handled in a call to SamrLookupNames. If the named pipe connection to the DC was reset in the middle of the SamrLookupNames call, lsass would crash.

The error mapping for STATUS_CONNECTION_RESET was not helpful, so it is switched so a better error message is returned from domainjoin.

CR Glenn

(lwbase: r61516)