MERGE: -c 61533 /trunk/Platform -> /branches/Platform-6.1
Fix bug 12971 - broken symlinks in pam.d cause domainjoin to fail

A broken symlink in the pam.d directory caused domain join to fail with a ERROR_FILE_NOT_FOUND message. The fix is to log a warning if a broken symlink is detected and skip that entry.

(domainjoin: r61534)