Fix bug 13145 - Domain join fails with an error "Error: WSAECONNRESET [code 0x00002746]"
This bug was a regression caused by revision 61515. The ERROR_NETNAME_DELETED code is marked as connection error, which means lsass will retry some calls that fail with that error.

Revision 61515 changed the error mapping of STATUS_CONNECTION_RESET to WSAECONNRESET, so WSAECONNRESET needs to be in the list too.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61536 9fc1c11e-11b2-4e14-94d8-6c5e05775969