Fix bug 13118 - domain join failed in Solaris 10 with "ERROR_FILE_NOT_FOUND" error
Lwio was crashing due to a double free of the GSS context. The bug affects other operating systems too, but the affects are most noticible on Solaris.

Freeing the tree connect's gss context is now done through RdrSessionSetupComplete2, so it does not need be done in the cleanup block of RdrNegotiateGssContextWorkItem2 and RdrNegotiateGssContextWorkItem.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61557 9fc1c11e-11b2-4e14-94d8-6c5e05775969