Fix bug 13169 - domain user password change fails if "user must change password at next logon" flag is set
When gss_init_sec_context is called for the spnego mechanism, it is ignoring all credentials acquired through other mechanisms. Requesting a spnego credential was actually acquiring all credential types. This meant with the standard Kerberos authentication, there was no way of restricting which mechanisms were used inside of spnego.

The 6.0 and 5.3 branches have a patch applied to Kerberos to pass the non-spnego credentials into spnego. This worked because the spnego credentials structure exactly matched the general gss structure. With Kerberos 1.8 and higher the structures no longer match up.

The new fix is to have spnego only acquire credentials for the other mechanisms that were requested at the same time.

Additionally, there was a dead lock in the ntlm credentials freeing code from lwio. The fix is to make NtlmFreeCredentialsHandle async.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@61680 9fc1c11e-11b2-4e14-94d8-6c5e05775969