Bug 13166 - [Mac10.7.2 -INTERNAL]LWISE : Password change failed for the domain user
10.7.2 is now doing its own password change handling based on the
value of the AuthenticationAuthority attribute, and the value we
have been returning, ";basic;", is not supported and causes an error.
We want it to just call into our DS plugin for password changes,
and the way to get it to do that is apparently to not return an
AuthenticationAuthority attribute at all.

CR: Kyle

(lsass: r61816)