MERGE: -c 61871 /trunk/Platform -> /branches/Platform-6.1
Fix build break

Do not link curl against libssh2. On Solaris i386, it is pulling in the headers in /opt/csw/include, but not the library. In either case, we should not pull in csw libraries. Disabling libssh2 suppport matches the behavior on Solaris from before libssh2 was installed on the build machine.

(curl: r61873)