Fix bug 13462 - unable to id users with nscd running
The code that edits the apparmor settings bitrot. It seems that domain join has not been editing the file for a while, but SuSE has been including some of the necessary Likewise settings by default.

Since the product was renamed, the SuSE default settings stopped working. Nscd was blocked by apparmor from opening the pbis libraries.

This change fixes editing the apparmor file.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@62997 9fc1c11e-11b2-4e14-94d8-6c5e05775969