Fix bug 13708 - assertion failed in Access Privilege Changes report
An assert failed in the Access Privilege Change report. This bug occurred because in the BVT ou in lampi.centeris.com, both LAMPI\lla and 'LAMPI\special user' have the alias 'special'.

The duplicate alias caused the user monitor to generate events of the form:
User special added
User special added
User special deleted
User special deleted

Slava's assertion tripped on a non-existent user being deleted (the second add was ignored).

The fix is for the user monitor to keep an in-memory hash of all the users and groups it has processed by name. When a duplicate alias is found, the second user is reported in NT4 format instead, and an error is sent to syslog.

If you are encountering this bug, you must delete all events in your enterprise database, then generate new events with a new build to get the fix.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@63729 9fc1c11e-11b2-4e14-94d8-6c5e05775969