Fix bug 13732 - names outside the basic multilingual plane are not handled by lsassd or usermonitor
Users with a character that cannot be encoded in UCS-2 cannot login through lsass, and they cause usermonitor to terminate.

The termination is because the registry cannot store their value as a string.

The fix for the usermonitor is to ignore these users. This meets the expectations of the usermonitor since it only lists users who can log in.

The registry and lwbase changes are to get a better error code from lwreg when the string conversion fails.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@63745 9fc1c11e-11b2-4e14-94d8-6c5e05775969