MERGE: Platform-6.5 -> Platform r64052 Bug 13939 - [All -INTERNAL] PBISE: Eventlog-cli is not displaying the event details properly.
Properly fix handling of log messages with a mix of ASCII and multibyte
sequences. The underlying problem is the length specifier of printf-family
of functions does not have useful semantics when applied with wchar_t
(%ls) format specifier. As a result, it is safer to rely on NULL-terminated
strings than using a length specifier.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@64057 9fc1c11e-11b2-4e14-94d8-6c5e05775969