MERGE: -c 63985 /trunk/Platform -> /branches/Platform-7.0
Fix bug 13846 - local and AD logins fail on HP-UX PA-RISC with latest version of sshd, 5.9 (64bit)

Create a wrapper function, LwStrtoll that calls the appropriate strtoll function on the platform. Use that wrapper in lwadvapi and lsass. Previously the nsswitch libraries were trying to load a non-existent strtoll function in 64bit mode.

The library constructor is not getting called in lwmsg to initialize xnet. The fix is to verify the library is loaded at the top of every wrapper function.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-7.0/src/linux@64069 9fc1c11e-11b2-4e14-94d8-6c5e05775969