MERGE: -c 64191 /trunk/Platform -> /branches/Platform-7.0
Partial fix for bug 13973 - lsass causes usermonitor daemon to die on Mac OS X 10.7.3

Lsass enumerated a duplicate user, but would not look it up by name. This caused usermonitor to treat the user as a local user, not AD user. Changing lsass's behavior would be too large of an API change.

This change allows user monitor to store local user names with slashes in them. So even though the AD user will be reported as a local user, usermonitor will not abort.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-7.0/src/linux@64192 9fc1c11e-11b2-4e14-94d8-6c5e05775969