Merge -c 63862 ^/branches/Platform-6.5 Fix bug 13469 - lsass missed domain local group membership for cross-forest user s
If the computer is joined to a different domain than the user whose membership i
s being looked up, search for all domain local groups in the computer's domain t
hat have either the user's DN or a foreign security principal for the user in th
e member attribute.

Searching for the foreign security principal DN is necessary for users from different forests. Searching for the user DN is necessary for users in the same forest (but different domain).

This checkin does not make domain local groups show up for users from a one-way trust. Only memberships for users from two-way trusts are shown. For one-way trusts, the existing behavior is still followed where only the primary group is shown until the user logs in via password.


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@64207 9fc1c11e-11b2-4e14-94d8-6c5e05775969