Patch 3 for bug 14600: Conflict Resolution in the packages
This may affect QA tests if those tests are not using the install.sh and those tests are not using "rpm -Uvh" for installation (U for upgrade, vs. "-i" for install.

On Debian/Ubuntu, this change will block raw installation via dpkg - this is by design and handled by the install.sh or via "apt-get" if the packages exist in a repository apt can search.

I expect no breaking of QA tests, since they should be using the install.sh.  But if it does, that's where to look first.

See: https://bt-bugzilla.pbis.beyondtrust.com/show_bug.cgi?id=14600



git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-7.1/src/linux@65547 9fc1c11e-11b2-4e14-94d8-6c5e05775969