Defect fix: registry delete tree was failing on 64 bit big endian platforms.  #53628
SqlDeleteTreeInternal_inlock_inDblock attempts to delete a tree by
recursively deleting subkeys. When obtaining the subkey count it was
incorrectly casting a DWORD (4 bytes) value to a size_t (8 bytes).

For a domainjoin leave this was preventing the deletion of the machine
account information resulting in the domainjoin cli thinking (briefly)
that the leave failed.