Stopping services uses a kill timer #46260.
Added a 'generic' simple pthread timer that runs a supplied action when
it expires. When an entry is stopped LwSwTableStopEntry() starts a timer
that kills the service if it doesn't shutdown before the timer expires.

This currently uses a hardcoded delay, but should use the same
configurable and default values for the service shutdown timer.