Merged PR 3974: Merge TFS116600-domainjoin-cli-history-support to master
Adds history support for parameters entered at linenoise prompts (excludes passwords). This applies to join, leave and setname.

Parameters are persisted in $HOME/.pbis-djname-hist and $HOME/.pbis-djjoin-hist for setname and join/leave respectively.

If $HOME is not set, this attempts to read the home directory from the effective user passwd entry.  If the home directory can't be determined, parameters will not be persisted between domainjoin-cli runs.

Related work items: #116600