Merged PR 6916: If --password=- read new-computer password from stdin #151748
If --password=- read new-computer password from stdin using the same mechanism as used for --name and --dn.
No password prompt is produced (which is typical when reading from stdin), and setting the log level to verbose or trace will show the password adtool is using for the computer object.

Related work items: #151748