Merged PR 7362: Merge defects/155710-API-On-Mac to master
On the mac building both 32 and 64. After building curl either 32 or 64, curl.h, curlbuild.h and curlrules.h are stage to stage/opt/pbis. When 64 is built it tries including the 32 bit version of the headers and the type size checking in curlrules.h fails. Change is to not include the headers filles in the stage directory, but the header files in object/curl/build-host-x86-64 or object/curl/build-host-x86-32.

Related work items: #155710