Merged PR 7387: Merge TFS121216-add-man-pages to master
Adds a pbis-open and domainjoin-cli man page.

The man pages are generated from markdown formatted files and processed via scripts to include/exclude enterprise sections and to support indenting of paragraphs.  The generated man pages are checked in so that they do not have to be generated during the build.

The overall intent is that the markdown formatted files should be easier to edit and maintain.

Related work items: #121216