Merged PR 7583: #154933-Support PasswordSafe password change upon requestId checkin
Currently domain join using PasswordSafe credentials involve requesting, then checkin the password. The password is subsequently used for the domain join. However, this may not work if PasswordSafe is configured to change the password upon checkin. The fix is to request the password from PS, do the domain join then checkin the password.

Related work items: #154933