Add option to control the computer account creation of domainjoin
PBIS creates a disabled computer account when domain join and move the computer
account and enable it when domain join successed. However it violate some
restrictions of some domain, so it will fail a domain join progress.

This patch add an option `--no-pre-account` to switch off the creation of the
disabled computer account when domain join.

Signed-off-by: Yang Qian <yang.qian@citrix.com>